refactor: 교통편 조회 API 모듈 책임 재분배
